rename enum to class_types
