+++ BALSAMIC/constants/analysis.py
@@ -1,5 +1,5 @@
 """Balsamic analysis specific constants."""
-from BALSAMIC.utils.str_enum import StrEnum
+from BALSAMIC.utils.class_types import StrEnum
 
 
 class ConfigType(StrEnum):
